Add int cast to capture stream value

+++ src/main/java/org/example/Classroom.java
@@ -38,7 +38,7 @@ public class Classroom {
     }
 
     public int getHeadCount(){
-        return Arrays.stream(this.students).filter(student -> student != null );
+        return (int) Arrays.stream(this.students).filter(student -> student != null);
     }
 
     public Classroom(int maxStudents) {
